Binaries that the CPython 3.7.2 build process depends on
As of CPython 3.7.2 the build process creates a "TCL_LIBRARY.env"
in the "$(OutDir)", which stores the full path of the {dde1.4,nmake,
reg1.3,tcl8,tcl8.6,tix8.4.3,tk8.6} TCL script directories, that are
shipped with a full Tcl/Tk release.

Therefore, this release of the "CPython89 dependencies" augments
the previously released, pre-built "tcltk-8.6.9.1" distribution
with all .tcl scripts that were previously missing.